# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Paper Company",
    "version": "1.0.0",
    "description": "Witaj w repozytorium Systemu ERP dla Papier Company - zainspirowanego przez kultowy serial \"The Office\". Nasz system jest zaprojektowany, aby wspierać i optymalizować kluczowe procesy biznesowe w firmie zajmującej się dystrybucją papieru, od zarządzania zapasami po finanse i księgowość."
  },
  "paths": {
...
    "/register": {
      "post": {
        "summary": "Rejestruje nowego użytkownika.",
        "description": "Endpoint do rejestracji nowego użytkownika w systemie.",
        "tags": [
          "Register"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "Adres email użytkownika."
                  },
                  "username": {
                    "type": "string",
                    "description": "Nazwa użytkownika."
                  },
                  "password": {
                    "type": "string",
                    "description": "Hasło użytkownika."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pomyślnie zarejestrowano użytkownika.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat potwierdzający rejestrację."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Błędne żądanie.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Adres email lub nazwa użytkownika już istnieją.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/register/{token}": {
      "get": {
        "summary": "Aktywuje konto użytkownika.",
        "description": "Endpoint do aktywacji konta użytkownika na podstawie tokena.",
        "tags": [
          "Register"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "description": "Token aktywacyjny.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Pomyślnie aktywowano konto użytkownika.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string",
                  "description": "Przekierowanie do strony logowania."
                }
              }
            }
          },
          "401": {
            "description": "Błąd uwierzytelniania.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "summary": "Pobiera listę wszystkich użytkowników.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano listę użytkowników.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "usersList": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Endpoint dostępny tylko dla użytkowników o roli \"admin\".",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": "// Odpowiednie pola, takie jak id, username, email itp."
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/user/{id}": {
      "get": {
        "summary": "Pobiera informacje o użytkowniku.",
        "description": "Pobiera informacje o użytkowniku na podstawie jego identyfikatora.",
        "tags": [
          "nn"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Identyfikator użytkownika.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pomyślnie pobrano informacje o użytkowniku.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "403": {
            "description": "Brak dostępu do zasobu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Błąd serwera.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Komunikat błędu."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}